requirements.txt
flask
python-docx
PyPDF2
nltk
numpy
scikit-learn
transformers
torch